requirements.txt
praw
python-dotenv
ollama 